Split CalculationValuesMap type into its own files https://bugs.webkit.org/show_bug.cgi?id=295536
Reviewed by Darin Adler.

This will let it be used from multiple places.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/Length.cpp:
* Source/WebCore/platform/calc/CalculationValueMap.cpp: Added.
* Source/WebCore/platform/calc/CalculationValueMap.h: Added.

Canonical link: https://commits.webkit.org/297111@main
